Add "logo"
I chose to create an image with the Rust logo plus something representing Objective-C; since they don't have a logo, I used the short form "Obj-C" enclosed in square brackets which is kind of a play on the message sending syntax in Objective-C. I've seen this style elsewhere, so no invention on my part.

The Rust logo is borrowed from here:
https://github.com/rust-lang/rust-artwork/blob/809592a8fafee0c70279d80905aba768aa1dd035/logo/rust-logo-blk.svg

Since this crate is not commercial, and is clearly related to (but not officially part of) the Rust project, usage of the trademark should be acceptable.