Merge objc2_id into objc2
Id is such a vital requirement for applications that there's no need to make it a "second class citizen". If we really feel it's too much of an experiment, we can always feature-flag it later.

Also, there's might be traits I'd like to implement for e.g. Option<Id> that is not possible if it lives in it's own crate!