Implement RefEncode for integer types
This was not done previously because of uncertainty with the `char*` being the "String" encoding.

But investigation shows that:
- @encode(signed char) -> "c"
- @encode(unsigned char) -> "C"
- @encode(char*) -> "*"
- @encode(char**) -> "^*"

And here it doesn't matter if we replace `char` with e.g. `uint8_t`, the encoding scheme still works like this.

Luckily `RefEncode` is designed in just the right way that it can actually handle this really nicely!