Set current version numbers
I've released a few intermediate versions of the crates as they looked while I worked on them. The intention is to make it easier for users to slowly migrate from `objc` to `objc2`.

These version bumps are not tracked in `git` (apart from tags), so this commit updates the currently used version numbers