Rename `links` key in sys crates
We're using the [`links`](https://doc.rust-lang.org/cargo/reference/build-scripts.html#the-links-manifest-key) key in `objc-sys` and `block-sys` to allow the user to customize linking, and use build script output in their own build scripts.

However, these are _global_, meaning that you can't have two packages with the same key - this goes for packages with different versions.

To avoid a situation where we can't use a new breaking version of these crates because everyone would have to update at the same time (no, the semver trick can't really save us here), we change the `links` key to include the version preemptively. Once we reach version 1.0 for these crates, we can go back to having a `links` key without a version number in it.