Merge block-sys into block2
`block-sys` doesn't actually build a library, it only links to it, so there is no real need for a build script, nor for the `links` key.

Additionally, it should actually only expose the six public symbols, and in that case, it just seems wasteful to have multiple crates.

This _does_ remove:
1. The ability to override the linking. This was brittle already, and is probably better handled by dedicated tools / by modifying the binaries after they're compiled / by inserting extra linker commands.
2. The `DEP_BLOCK_0_2_CC_ARGS` environment variable exposed to downstream build scripts. This was not really that useful, usually you want to know exactly what flags you're passing to your compiler, and the standard `-fblocks` flag is enough.