m: Optmize is_main_thread by using a libc call
Previous, the "is_main_thread" function would always make a call to the
"NSThread::isMainThread_class" function to determine if the current thread is
the main thread. This requires an objc dispatch call every time it is called. As
this function or MainThreadMarker is used frequently throughout the objc2
ecosystem it takes a lot of time.

The NSThread function just calls libc::pthread_main_np under the hood. Benchmarks
show that calling the libc function instead is four times faster. So on platforms
that have it we just call that instead.

Signed-off-by: John Nunley <dev@notgull.net>