Make getting references bound by the autorelease pool `unsafe`
The current design of `autoreleasepool` does not prevent the user from
"smuggling" the lifetime from an outer pool and using that inside an
inner pool, even with the unstable auto traits guard.

So let's make using the pool to construct a reference `unsafe` (making
the invocation of `autoreleasepool` itself `unsafe` would be cumbersome
for a lot of applications, especially so because of the extra nesting).

Fixes https://github.com/madsmtm/objc2/issues/540, see that for details
on why this is unsound.

We _might_ be able to make this pattern safe again with something like
contexts and capabilities, but that's future work.