Make generic container types InteriorMutable
Affects NSEnumerator, NSArray, NSDictionary, NSSet and the mutable
variants.

This removes all `_mut` methods and `Mut` iterators, as these are now
both unsound and unnecessary.

Most methods are also converted to `_retained` variants, as returning
references is unsound if the collection is mutated. `_unchecked`
variants are provided as alternatives instead.

At the same time, we remove a bunch of helper methods that are now
unnecessary, including `from_vec`.

Finally, we add a fuzz target for ensuring that the safe methods do
catch mutation mistakes.