Forward implementations of mutable traits on `Retained<T>` to `&T`
This makes it possible to still implement such traits that require
mutability, even though the type itself uses interior mutability.

The method through which this is accomplished differs a bit depending on
the specific trait, but generally we try to implement mutable traits
for both `Retained<T>` and `&Retained<T>`.

Part of https://github.com/madsmtm/objc2/issues/563