Completely remove the ability to create mutable objects
The ability to specify whether an object is thread-safe or not has been
kept, `ClassType::Mutability` has been renamed to `ThreadKind` and used
in that way instead. It can be omitted from `extern_class!` most of the
time, since we can make it default to that of the superclass.

Finally fixes https://github.com/madsmtm/objc2/issues/563