Refactor how NSCopying return types are determined
Previously, we used `CounterpartOrSelf`, which was a nice hack to work
around Rust not having associated type defaults.

Now, we use the user-facing `objc2_foundation::[Mutable]CopyingHelper`,
which are intended to be implemented by the user alongside `NSCopying`.

This is more verbose for the common case where the user just wants to
say that something implements `NSCopying`, but is also better separation
of concerns, and allows us to simplify `ClassType` further.

We could consider a `#[derive(NSCopying)]` that does this, see:
https://github.com/madsmtm/objc2/issues/267

Part of https://github.com/madsmtm/objc2/issues/563