Use the same generic name in all collection impls
We previously used names such as `T`, `K` and `V`, which is the common
Rust style. These generics aren't used in the generated interface
though, which might cause confusion, so let's use the same generic name
everywhere.

In short, the change is:
- `NSArray<T>` -> `NSArray<ObjectType>`
- `NSSet<T>` -> `NSSet<ObjectType>`
- `NSDictionary<K, V>` -> `NSDictionary<KeyType, ObjectType>`

I was unsure about the helper `Q` generic on `NSDictionary`, I went with
calling that `CopiedKey`.