Change extern_class! syntax
Instead of specifying `impl ClassType for ...`, we instead parse the
custom attributes `#[unsafe(super(...))]`, `#[thread_kind = ...]` and
`#[name = ...]`.

This is nice because:
- It's more concise.
- It more closely matches what we might end up with once it can become
  and attribute macro:
  https://github.com/rust-lang/rfcs/pull/3697
- We need to parse the attributes anyhow to override derives:
  https://github.com/madsmtm/objc2/issues/267

  (The extern_class! part of that issue is now resolved).
- It makes it easier to change ClassType in the future without having to
  change the macro API as well.

Additionally, this commit also adds incomplete support for generics, to
avoid the framework crates depending on an internal macro, and it
improves rust-analyzer support in extern_class! by having more relaxed
parsing.