Use Clang's -fmodules
Clang has something called "modules", which are a mechanism to describe
and combine a set of headers into a single cohesive unit:
https://clang.llvm.org/docs/Modules.html

This is what Swift uses when mapping frameworks, so it makes sense for
us to do the same. Objective-C also recently gained the `@import` syntax
to do the same.

This should unblock our support for umbrella frameworks, makes it easier
to use API notes in the future, and is generally more correct.