Generate encoding tests
This found at least one mistake, tensorByDequantizingToType:scale:bias:
in MLCompute, which seems to take a float internally that it doesn't
declare in the header! It also found that NSDecimal had the wrong
encoding.

This might also allow us to remove encoding testing at runtime from
`extern_methods!` in the future.