Update assembly tests
Missed in 9c74cc3. Outside of the unwind info which changes wildly (as
expected), there is only a single functional change, namely that when
using `Option<&mut ...>` now has a single main call to `objc_msgSend,
instead of the early branch it had before that just forwarded to
`objc_msgSend`.

I'm not entirely sure why LLVM decides to do this now, I'm guessing it
has something to do with the extra unwind info possibly inhibiting the
previous optimization.

In practice, I doubt it will make much of a difference, as passing
`None` to to these is rarely what you want (you are often interested in
the out parameter).