Push impl inside extern_methods!
It was previously required for us to have the `unsafe impl` inside
`extern_methods!`, since we had to have _somewhere_ to write the
`unsafe` keyword.

Since 227da1e, the keyword is now part of the #[method] attribute, and
so it is now redundant at the `impl` level.

So let's just completely remove the `impl` nested inside
`extern_methods!`, and instead nest _that_ inside `impl`.