Rename dispatch2 types
Prefix them with "Dispatch", to better match Swift, and to allow easily
glob-importing from dispatch2 without fear that the symbols will clash
with something else.

This isn't really "rusty", but it plays much nicer with the rest of the
objc2 ecosystem, and matches what Swift does.

It's also clearer at the usage-site. Compare:

   DispatchQueue::main().exec_sync(|| ...)

And:

   Queue::main().exec_sync(|| ...)