Prepare block2 for future breaking changes
I plan to convert `Block<dyn Fn() + '_>` to `Block<'_, fn()>` to allow
for parameters with lifetimes (the Fn trait is invariant, while fn()
types have variance).

Doing this will require breaking changes in block2, so to prepare for
that, we introduce the helper `DynBlock`, which should allow converting
`DynBlock<dyn Fn() + '_>` to `Block<'_, fn()>` in the next version.