Convert CoreFoundation-like functions to methods
This is important for ergonomics, it's a lot of hazzle to write out the
entire function, and doesn't communicate information to always have to
put an & in front of retained arguments.

This is also better for discoverability (it's easier to see which
methods are available on a given type in the docs).

Fixes https://github.com/madsmtm/objc2/issues/736.

I've had to rename a few methods in objc2-core-foundation, but
fortunately we haven't published those yet, so that's fine.