Improve logic for converting functions into method
We previously preferentially tried first types that the function
mentioned in it's argument and return types. Now, we instead assemble a
list of "candidates" for the method, where the argument and return types
are (mostly) ignored.

This also allows us to use `self: *` in a few more places, since before,
we didn't correctly handle fake typedefs.