Rename some Dispatch types to match Swift
DispatchTime, DispatchQoS, DispatchAutoReleaseFrequency, DispatchBlock,
and a few DispatchIO* types.

At the same time, remove the `dispatch2::ffi` module, it is now mostly
redundant, and it's a bad idea to split the API in two like this: we
should instead work towards mapping everything, so that there isn't a
need!

Part of https://github.com/madsmtm/objc2/issues/77.