Add AsRef impls for toll-free bridging
I may have went a little overboard with this, considering that it's only
CoreFoundation, CoreText and IOSurface that actually _have_ bridged
types...

But a lot of that complexity was from improvements to macro parsing,
which might be useful later, so we'll go with it for now.

Fixes https://github.com/madsmtm/objc2/issues/693.