Add OpenGL, OpenGLES and GLKit
These are deprecated by Apple, but are still used in a few places (such
as Servo, but probably many more applications in the wild), and hence
still make sense for us to emit bindings for.

I have intentionally not mapped the entirety of the OpenGL API, only the
Apple-specific parts that they provide extra headers for.

Fixes https://github.com/madsmtm/objc2/issues/722.