Refactor safety checking
To be able to return another state for when safety is unknown, and to
allow differentiating between argument and return safety.

The latter is needed for blocks, which can actually be unsafe to return
if they take a pointer as an argument (because in that case, we'd return
the block safely, and allow the user to call it with a bogus pointer).

We also disallow AnyObject/CFType for now, see:
https://github.com/madsmtm/objc2/issues/562