Generate safety comments
Both when we're sure that something is unsafe, and when we're uncertain.

The safety comment might of course not be complete, there might be
further requirements that you have to uphold for API usage to be sound
that we can't automatically detect. This is a bit unfortunate, since the
mere presence of the "Safety" tag might lead users to thinking that
they've done all they need to do.

I'm okay with that tradeoff, I'd rather have these comments than not.