Avoid recompiling `objc2` when `objc2-foundation` changes
`objc2` previously had a dev dependency on `objc2-foundation`, to allow
doc tests to reference types from that. I did this back in eb39164 to
improve the documentation, and while it's a nice idea, it is troublesome
for iteration time, since if either `objc2` or `objc2-foundation`
changes, Cargo has to rebuild both before it can run any tests.

I think a better approach is to use `#[path = ...]` and run the doctests
in the `tests` crate instead.