Fix various definitions on Mac Catalyst and the Simulator
Fixed APIs:
- SSLCipherSuite.
- CMBaseClassVersion and CMStructVersion.
- MTLCopyAllDevices.
- CSSM_TP_APPLE_EVIDENCE_INFO.
- UIImageResizingMode, NSImageResizingMode and NSTextAlignment consts.
- AudioSampleType and AudioUnitSampleType.
- kAudioFormatFlagsCanonical and kAudioFormatFlagsAudioUnitCanonical.

Found by searching for TARGET_OS_MACCATALYST and TARGET_OS_SIMULATOR.

This uses `target_env` instead of `target_abi`, since that's in our
MSRV, although in a bit of a weird way: If we used `target_abi`, we'd be
able to have an MSRV of 1.78, while if we use `target_env`, we can have
a lower MSRV for most targets, but for Mac Catalyst / Simulator targets,
the MSRV is raised to 1.91. This isn't ideal, but it's the best we can
do without a build script.