Move ivars back into define_class!
This has a better `Debug` impl, is shorter, and is more in line with
what the user would expect.

It also allows us to extend it in the future with an attribute like
`#[name = "my_ivar"]` on the field to make the ivar publicly visible on
the class (which is important for KVO), and maybe even at some point
a `#[property(getter, setter = setIvar)]` attribute for generating
Objective-C getter and setter methods.

This fixes https://github.com/madsmtm/objc2/issues/788.

A downside is that deriving `Default` is no longer possible, which means
that some `init` methods are a bit longer, but we should really solve
that by allowing some sort of `Default` on the class itself, see
https://github.com/madsmtm/objc2/issues/267.